:recycle: refactor: Change ID type from Long to Integer

+++ src/main/java/excluz/excluz/domain/cartItem/repository/CartItemRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import excluz.excluz.common.entity.CartItem;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
 }
